+++ autogpt/agent/__init__.py
@@ -1,4 +1,4 @@
-from autogpt.agent.agent import Agent
-from autogpt.agent.agent_manager import AgentManager
+from .agent import Agent
+from .agent_manager import AgentManager
 
 __all__ = ["Agent", "AgentManager"]
